fix: Alter type id to long

+++ Desafio.BackEnd/Interfaces/IRepository.cs
@@ -3,8 +3,8 @@
     public interface IRepository<T> where T : class
     {
         public Task<List<T>> GetAllAsync();
-        public Task<T> GetByIdAsync(int id);
-        public Task DeleteAsync(int id);
+        public Task<T> GetByIdAsync(long id);
+        public Task DeleteAsync(long id);
         public void Update(T entity);
         public Task<T> CreateAsync(T entity);
     }
